hittarget: Move updater code into separate class.
